Move runAction function to it's own file
